PR binutils/13135 	* arm-dis.c: Add necessary casts for printing integer values. 	Use %s when printing string values. 	* hppa-dis.c: Likewise. 	* m68k-dis.c: Likewise. 	* microblaze-dis.c: Likewise. 	* mips-dis.c: Likewise. 	* ppc-dis.c: Likewise. 	* sparc-dis.c: Likewise.
	* dis-asm.h (fprintf_ftype): Add ATTRIBUTE_FPTR_PRINTF_2.

+++ opcodes/ppc-dis.c
@@ -666,7 +666,7 @@ print_insn_powerpc (bfd_vma memaddr,
 	      (*info->fprintf_func) (info->stream, "%s", cbnames[cc]);
 	    }
 	  else
-	    (*info->fprintf_func) (info->stream, "%d", value);
+	    (*info->fprintf_func) (info->stream, "%d", (int) value);
 
 	  if (need_paren)
 	    {
